requirements.txt
black
icecream
isort
matplotlib
numpy
pandas
seaborn
scikit-learn
scipy
torch
torchvision
tqdm
py-sudoku
ucimlrepo
gputil
natsort
gillespy2
ucimlrepo
langevin-sampling